:white_check_mark: [#1280] Test copying form
